+++ spring-web/src/test/java/org/springframework/web/service/invoker/RequestPartArgumentResolverTests.java
@@ -23,9 +23,12 @@ import reactor.core.publisher.Mono;
 
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.bind.annotation.RequestPart;
+import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.service.annotation.PostExchange;
+import org.springframework.web.testfixture.servlet.MockMultipartFile;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -46,6 +49,9 @@ class RequestPartArgumentResolverTests {
 
 	private final Service service =
 			HttpServiceProxyFactory.builderFor(this.client).build().createClient(Service.class);
+
+	private static final MockMultipartFile mockMultipartFile = new MockMultipartFile(
+			"testFileName", "originalTestFileName", "text/plain", "test".getBytes());
 
 
 	// Base class functionality should be tested in NamedValueArgumentResolverTests.
@@ -69,6 +75,54 @@ class RequestPartArgumentResolverTests {
 		assertThat(map.getFirst("optionalPart").getBody()).isEqualTo("part 4");
 	}
 
+	@Test
+	void multipartFile() {
+		this.service.postMultipartFile(mockMultipartFile);
+		testMultipartFile(mockMultipartFile, "file");
+	}
+
+	@Test
+	void requestPartMultipartFile() {
+		this.service.postRequestPartMultipartFile(mockMultipartFile);
+		testMultipartFile(mockMultipartFile, "myFile");
+	}
+
+	@Test
+	void requestPartOptionalMultipartFile() {
+		this.service.postRequestPartOptionalMultipartFile(Optional.of(mockMultipartFile));
+		testMultipartFile(mockMultipartFile, "file");
+	}
+
+	@Test
+	void optionalMultipartFile() {
+		this.service.postOptionalMultipartFile(Optional.empty(), "anotherPart");
+		Object value = client.getRequestValues().getBodyValue();
+
+		assertThat(value).isInstanceOf(MultiValueMap.class);
+		@SuppressWarnings("unchecked")
+		MultiValueMap<String, HttpEntity<?>> map = (MultiValueMap<String, HttpEntity<?>>) value;
+		assertThat(map).containsOnlyKeys("anotherPart");
+	}
+
+	private void testMultipartFile(MultipartFile testFile, String partName) {
+		Object value = this.client.getRequestValues().getBodyValue();
+
+		assertThat(value).isInstanceOf(MultiValueMap.class);
+		@SuppressWarnings("unchecked")
+		MultiValueMap<String, HttpEntity<?>> map = (MultiValueMap<String, HttpEntity<?>>) value;
+		assertThat(map).hasSize(1);
+
+		HttpEntity<?> entity = map.getFirst(partName);
+		assertThat(entity).isNotNull();
+		assertThat(entity.getBody()).isEqualTo(testFile.getResource());
+
+		HttpHeaders headers = entity.getHeaders();
+		assertThat(headers.getContentType()).isEqualTo(MediaType.TEXT_PLAIN);
+		assertThat(headers.getContentDisposition().getType()).isEqualTo("form-data");
+		assertThat(headers.getContentDisposition().getName()).isEqualTo(partName);
+		assertThat(headers.getContentDisposition().getFilename()).isEqualTo(testFile.getOriginalFilename());
+	}
+
 
 	private interface Service {
 
@@ -78,6 +132,17 @@ class RequestPartArgumentResolverTests {
 				@RequestPart Mono<String> part3,
 				@RequestPart Optional<String> optionalPart);
 
+		@PostExchange
+		void postMultipartFile(MultipartFile file);
+
+		@PostExchange
+		void postRequestPartMultipartFile(@RequestPart(name = "myFile") MultipartFile file);
+
+		@PostExchange
+		void postRequestPartOptionalMultipartFile(@RequestPart Optional<MultipartFile> file);
+
+		@PostExchange
+		void postOptionalMultipartFile(Optional<MultipartFile> file, @RequestPart String anotherPart);
 	}
 
 }
